Add some missing components to barrel files and rename component files
