+++ src/components/Layout/ContentSection/SplitContent.tsx
