+++ src/components/Layout/ContentSection/TextWithImage.tsx
