+++ src/components/Layout/index.ts
@@ -2,3 +2,4 @@ export * from './Navigation/Navigation';
 export * from './Footer/Footer';
 export * from './Layout';
 export * from './LayoutArticle';
+export * from './Logo';
