+++ src/components/Map/RsvMap.tsx
@@ -4,8 +4,7 @@ import maplibregl, { LngLatBoundsLike } from 'maplibre-gl';
 import { StaticImage } from 'gatsby-plugin-image';
 import 'maplibre-gl/dist/maplibre-gl.css';
 import { getOptInCookie, OptIn } from '~/components/CookieConsent/';
-import { RSVSegment } from './RSVSegment';
-import { RSVPopup } from './RSVPopup';
+import { RSVSegment, RSVPopup } from '.';
 
 type Props = {
   geometry: {
@@ -40,7 +39,7 @@ export const RSVMap: React.FC<Props> = ({ geometry }) => {
     return (
       <div className="relative max-h-full max-w-full">
         {consent === false && (
-          <div className="absolute top-1/2 z-10 mx-2 -translate-y-1/2">
+          <div className="absolute top-1/2 z-10 mx-2 -translate-y-1/2 md:mx-5">
             <OptIn setConsent={setConsent} />
           </div>
         )}
