+++ src/components/Map/index.ts
@@ -1,1 +1,3 @@
-export * from './RSVMap';
+export * from './RsvMap';
+export * from './RsvPopup';
+export * from './RsvSegment';
